chore[github]: extract diff logic
